+++ src/sst/elements/vanadis/rocc/vbasicrocc.h
@@ -1,8 +1,8 @@
-// Copyright 2009-2023 NTESS. Under the terms
+// Copyright 2009-2025 NTESS. Under the terms
 // of Contract DE-NA0003525 with NTESS, the U.S.
 // Government retains certain rights in this software.
 //
-// Copyright (c) 2009-2023, NTESS
+// Copyright (c) 2009-2025, NTESS
 // All rights reserved.
 //
 // Portions are copyright of other developers:
@@ -40,20 +40,36 @@ public:
                                           "Implements a basic example of a RoCC accelerator",
                                           SST::Vanadis::VanadisRoCCInterface)
 
-    SST_ELI_DOCUMENT_PARAMS(
-            { "verbose", "Set the verbosity of output for the RoCC", "0" }, 
-            { "max_instructions", "Set the maximum number of RoCC instructions permitted in the queue", "8" },
-            {"clock",              "Clock frequency for component TimeConverter. MMIOTile is Unclocked but subcomponents use the TimeConverter", "1Ghz"}
-        )
+    SST_ELI_DOCUMENT_SUBCOMPONENT_SLOTS({ "memory_interface", "Set the interface to memory",
+                                            "SST::Interfaces::StandardMem" })
 
-    SST_ELI_DOCUMENT_STATISTICS({ "roccs_issued", "Count number of rocc instructions that are issued", "operations", 1 })
+    SST_ELI_DOCUMENT_PARAMS(
+            { "max_instructions", "Set the maximum number of RoCC instructions permitted in the queue", "8" },
+            { "clock",            "Clock frequency for component TimeConverter. MMIOTile is Unclocked but subcomponents use the TimeConverter", "1Ghz"}
+        )
 
     VanadisRoCCBasic(ComponentId_t id, Params& params) : VanadisRoCCInterface(id, params),
         max_instructions(params.find<size_t>("max_instructions", 8)) {
         
-        stat_roccs_issued = registerStatistic<uint64_t>("roccs_issued", "1");
+        try {
+            UnitAlgebra clock = params.find<UnitAlgebra>("clock", "1GHz");
 
-        UnitAlgebra clock = params.find<UnitAlgebra>("clock", "1GHz");
+            if (!(clock.hasUnits("Hz") || clock.hasUnits("s")) || 
+                clock.getRoundedValue() <= 0) {
+                output->fatal(CALL_INFO, -1,
+                    "%s, Error - Invalid param: clock.\n"
+                    "Must have units of Hz or s and be > 0.\n"
+                    "SI prefixes ok. You specified '%s'\n",
+                    getName().c_str(), clock.toString().c_str());
+            }
+        } catch (const UnitAlgebra::UnitAlgebraException& exc) {
+            output->fatal(CALL_INFO, -1,
+                "%s, Invalid param: Exception while parsing 'clock'.\n"
+                "'%s'\n",
+                getName().c_str(), exc.what());
+        }
+
+        std_mem_handlers = new StandardMemHandlers(this, output);
 
         busy = false;
         curr_resp = NULL;
@@ -62,6 +78,12 @@ public:
             "memory_interface", ComponentInfo::SHARE_PORTS | ComponentInfo::INSERT_STATS, getTimeConverter("1ps"),
             new StandardMem::Handler<SST::Vanadis::VanadisRoCCBasic>(
                 this, &VanadisRoCCBasic::processIncomingDataCacheEvent));
+
+        if ( nullptr == memInterface ) {
+            output->fatal(
+                CALL_INFO, -1,
+                "Error: unable to load memory interface subcomponent for VanadisRoCCBasic.\n");
+        }
 
     }
 
@@ -79,7 +101,7 @@ public:
     size_t roccQueueSize() override { return roccCmd_q.size(); }
 
     void push(RoCCCommand* rocc_me) override {
-        stat_roccs_issued->addData(1);
+        stat_rocc_issued->addData(1);
         roccCmd_q.push_back( rocc_me );
     }
 
@@ -90,7 +112,10 @@ public:
     }
 
     // initialize subcomponents and parameterizable data structures
-    void init(unsigned int phase) override {}
+    void init(unsigned int phase) override {
+        memInterface->setMemoryMappedAddressRegion(0, 64);
+        memInterface->init(phase);
+    }
 
     void tick(uint64_t cycle) override {
         output->verbose(CALL_INFO, 16, 0, "-> tick RoCC at cycle %" PRIu64 "\n", cycle);
@@ -104,31 +129,31 @@ public:
             busy = true; // start processing something
             curr_cmd = roccCmd_q.front(); // grab the command to process
 
-            output->verbose(CALL_INFO, 2, 0, "decoding func7 of RoCC inst\n");
+            output->verbose(CALL_INFO, 9, 0, "decoding func7 of RoCC inst\n");
             switch (curr_cmd->inst->func7) { 
                 case 0x0:
                 {
-                    output->verbose(CALL_INFO, 2, 0, "performing RoCC ADD\n");
+                    output->verbose(CALL_INFO, 9, 0, "performing RoCC ADD\n");
                     performADD();
                 } break;
                 case 0x1:
                 {
-                    output->verbose(CALL_INFO, 2, 0, "performing RoCC SRAI\n");
+                    output->verbose(CALL_INFO, 9, 0, "performing RoCC SRAI\n");
                     performSRAI();
                 } break;
                 case 0x2: // basic load
                 {
-                    output->verbose(CALL_INFO, 2, 0, "issuing load\n");
+                    output->verbose(CALL_INFO, 9, 0, "issuing load\n");
                     issueLoad();
                 } break;
                 case 0x3: // basic store
                 {
-                    output->verbose(CALL_INFO, 2, 0, "issuing store\n");
+                    output->verbose(CALL_INFO, 9, 0, "issuing store\n");
                     issueStore();
                 } break;
                 default: 
                 {
-                    output->verbose(CALL_INFO, 2, 0, "ERROR: unrecognized RoCC func7\n");
+                    output->verbose(CALL_INFO, 9, 0, "ERROR: unrecognized RoCC func7\n");
                 } break;
             }
         }
@@ -136,8 +161,10 @@ public:
 
     // adds rs1 and rs2 and writes it to rd
     void performADD() {
-        output->verbose(CALL_INFO, 2, 0, "EXECUTE RoCC ADD w/ rs1: %llx, rs2: %llx, result: %llx", curr_cmd->rs1, curr_cmd->rs2, curr_cmd->rs1 + curr_cmd->rs2);
-        completeRoCC(curr_cmd->rs1 + curr_cmd->rs2);
+        uint64_t rs1 = curr_cmd->rs1;
+        uint64_t rs2 = curr_cmd->rs2;
+        output->verbose(CALL_INFO, 9, 0, "EXECUTE RoCC ADD w/ rs1: %" PRIx64 ", rs2: %" PRIx64 ", result: %" PRIx64 "\n", rs1, rs2, rs1 + rs2);
+        completeRoCC(rs1 + rs2);
         return;
     }
 
@@ -146,7 +173,7 @@ public:
         uint64_t src_1 = curr_cmd->rs1;
         uint64_t shamt = curr_cmd->rs2;
         uint64_t result = src_1 >> shamt;
-        output->verbose(CALL_INFO, 2, 0, "EXECUTE RoCC SRAI w/ rs1: %llx, shamt: %llx, result: %llx + %llx", src_1, shamt, result);
+        output->verbose(CALL_INFO, 9, 0, "EXECUTE RoCC SRAI w/ rs1: %llx, shamt: %llx, result: %llx\n", src_1, shamt, result);
         completeRoCC(result);
         return;
     }
@@ -155,9 +182,8 @@ public:
     void issueLoad() {
         StandardMem::Request* load_req = nullptr;
 
-        load_req = new StandardMem::Read(curr_cmd->rs1, 4, 0, curr_cmd->rs1, 0, 0);
-        output->verbose(CALL_INFO, 9, 0, "----> Read Req: physAddr: %llx, size: %llx, vAddr: %llx, inst ptr: %llx, tid: %llx\n", 
-                        curr_cmd->rs1, 4, curr_cmd->rs1, 0, 0);
+        load_req = new StandardMem::Read(curr_cmd->rs1, 4, 0);
+        output->verbose(CALL_INFO, 9, 0, "----> Read Req: physAddr: %llx, size: %llx, vAddr: %llx, inst ptr: %llx, tid: %llx\n", curr_cmd->rs1, 4, curr_cmd->rs1, 0, 0);
 
         assert(load_req != nullptr);
 
@@ -168,6 +194,9 @@ public:
     void issueStore() {
         StandardMem::Request* store_req = nullptr;
         std::vector<uint8_t> payload(4);
+        for (int i = 0; i < 4; ++i) {
+            payload[i] = (rocc_internal_register >> (i * 8)) & 0xFF;
+        }
         
         store_req = new StandardMem::Write(curr_cmd->rs1, 4, payload, 
             false, 0, curr_cmd->rs1, 0, 0);
@@ -179,7 +208,7 @@ public:
     // turn off busy bit so host knows this accelerator is no longer busy
     // send RoCC response back to host
     void completeRoCC(uint64_t rd_val) {
-        output->verbose(CALL_INFO, 2, 0, "Finalize RoCC command w/ rd %" PRIu16 ", rd_val %" PRIu64 " \n", curr_cmd->inst->rd, rd_val);
+        output->verbose(CALL_INFO, 9, 0, "Finalize RoCC command w/ rd %" PRIu16 ", rd_val %" PRIu64 " \n", curr_cmd->inst->rd, rd_val);
         roccCmd_q.pop_front();
         busy = false;
         curr_resp = new RoCCResponse(curr_cmd->inst->rd, rd_val);
@@ -195,12 +224,12 @@ public:
         virtual ~StandardMemHandlers() {}
 
         virtual void handle(StandardMem::ReadResp* ev) {
-            out->verbose(CALL_INFO, 2, 0, "-> handle read-response (virt-addr: 0x%llx)\n", ev->vAddr);
+            out->verbose(CALL_INFO, 9, 0, "-> handle read-response (virt-addr: 0x%llx)\n", ev->vAddr);
             RoCCCommand* rocc_cmd = rocc->curr_cmd; // need to grab the instruction that generated the read request
             // so that we know where to store the read response results
 
             if ( ev->getFail() ) { 
-                out->verbose(CALL_INFO, 2, 0, "RoCC load failed, sending error code 1\n");
+                out->verbose(CALL_INFO, 9, 0, "RoCC load failed, sending error code 1\n");
                 rocc->completeRoCC(1);
             }
             
@@ -214,24 +243,30 @@ public:
                 register_value.at(reg_offset + addr_offset + i) = ev->data[i];
             }
 
-            switch (ev->getAllFlags()) { // treat read response data differently based on who issued it (flags)
-                default:
+            switch (ev->getAllFlags()) { // treat read response data differently based on who issued it
+            case 0x0: 
                 {
-                    rocc->output->verbose(CALL_INFO, 2, 0, "ERROR: unrecognized read response flag\n");
+                    rocc->rocc_internal_register = rocc->dataToInt(&register_value);
+                    rocc->output->verbose(CALL_INFO, 9, 0, "RoCC loaded value: %d\n", rocc->rocc_internal_register);
+                    rocc->completeRoCC(0); // complete the load command
+                } break;
+
+            default:
+                {
+                    rocc->output->verbose(CALL_INFO, 9, 0, "ERROR: unrecognized read response flag\n");
+                    rocc->completeRoCC(1); // complete the load command
                 } break;
             }
-
-            if (rocc_cmd->inst->func7 == 0x1) { // if this is just an individual load command:
-                rocc->completeRoCC(rocc->dataToInt(&register_value)); // complete the load command
-            }
+            
+            delete ev;
         } 
         
         virtual void handle(StandardMem::WriteResp* ev) {
             // write is much simpler because we aren't handling any reponse data
             // just need to make sure it went through properly
-            out->verbose(CALL_INFO, 2, 0, "-> handle write-response (virt-addr: 0x%llx)\n", ev->vAddr);
+            out->verbose(CALL_INFO, 9, 0, "-> handle write-response (virt-addr: 0x%llx)\n", ev->vAddr);
             if ( ev->getFail() ) { 
-                out->verbose(CALL_INFO, 2, 0, "RoCC store failed, responding with error code 1\n");
+                out->verbose(CALL_INFO, 9, 0, "RoCC store failed, responding with error code 1\n");
                 rocc->completeRoCC(1);
             }
             
@@ -244,13 +279,13 @@ public:
     };
 
     void processIncomingDataCacheEvent(StandardMem::Request* ev) {
-        output->verbose(CALL_INFO, 2, 0, "received incoming data cache request -> processIncomingDataCacheEvent()\n");
+        output->verbose(CALL_INFO, 9, 0, "received incoming data cache request -> processIncomingDataCacheEvent()\n");
 
         assert(ev != nullptr);
         assert(std_mem_handlers != nullptr);
 
         ev->handle(std_mem_handlers);
-        output->verbose(CALL_INFO, 2, 0, "completed pass off to incoming handlers\n");
+        output->verbose(CALL_INFO, 9, 0, "completed pass off to incoming handlers\n");
     }
 
     // takes the vector of 8bit chunks and converts it into single uint64_t value
@@ -274,8 +309,7 @@ public:
     StandardMem* memInterface;
 
     int max_instructions;
-
-    Statistic<uint64_t>* stat_roccs_issued;
+    int rocc_internal_register;
 };
 
 } // namespace Vanadis
