--- src/sst/elements/vanadis/tests/small/basic-io/hello-world-cpp-rocc/hello-world-cpp.cc
@@ -1,22 +1,0 @@
-// Copyright 2009-2023 NTESS. Under the terms
-// of Contract DE-NA0003525 with NTESS, the U.S.
-// Government retains certain rights in this software.
-//
-// Copyright (c) 2009-2023, NTESS
-// All rights reserved.
-//
-// Portions are copyright of other developers:
-// See the file CONTRIBUTORS.TXT in the top level directory
-// of the distribution for more information.
-//
-// This file is part of the SST software package. For license
-// information, see the LICENSE file in the top level directory of the
-// distribution.
-
-#include <iostream>
-
-int main(int argc, char* argv[]) {
-	std::cout << "Hello World from C++" << std::endl;
-
-	return 0;
-}
