--- src/sst/elements/vanadis/tests/small/basic-io/hello-world-cpp-rocc/riscv64/sst.stdout.gold
@@ -1,39 +1,0 @@
-pid=100 tid=100 has exited
-all process have exited
- node0.cpu0:decoder0:branch_unit.branch_cache_hit.1 : Accumulator : Sum.u64 = 16724; SumSQ.u64 = 16724; Count.u64 = 16724; Min.u64 = 1; Max.u64 = 1; 
- node0.cpu0:decoder0:branch_unit.branch_cache_miss.1 : Accumulator : Sum.u64 = 20478; SumSQ.u64 = 20478; Count.u64 = 20478; Min.u64 = 1; Max.u64 = 1; 
- node0.cpu0:decoder0:branch_unit.branch_cache_castout.1 : Accumulator : Sum.u64 = 3170; SumSQ.u64 = 3170; Count.u64 = 3170; Min.u64 = 1; Max.u64 = 1; 
- node0.cpu0:decoder0.uop_cache_hit.1 : Accumulator : Sum.u64 = 5386390; SumSQ.u64 = 5386390; Count.u64 = 5386390; Min.u64 = 1; Max.u64 = 1; 
- node0.cpu0:decoder0.predecode_cache_hit.1 : Accumulator : Sum.u64 = 13637; SumSQ.u64 = 13637; Count.u64 = 13637; Min.u64 = 1; Max.u64 = 1; 
- node0.cpu0:decoder0.predecode_cache_miss.1 : Accumulator : Sum.u64 = 5360815; SumSQ.u64 = 5360815; Count.u64 = 5360815; Min.u64 = 1; Max.u64 = 1; 
- node0.cpu0:decoder0.uops_generated.1 : Accumulator : Sum.u64 = 0; SumSQ.u64 = 0; Count.u64 = 0; Min.u64 = 0; Max.u64 = 0; 
- node0.cpu0:decoder0.decode_faults.1 : Accumulator : Sum.u64 = 0; SumSQ.u64 = 0; Count.u64 = 0; Min.u64 = 0; Max.u64 = 0; 
- node0.cpu0:decoder0.ins_bytes_loaded.1 : Accumulator : Sum.u64 = 21443260; SumSQ.u64 = 85773040; Count.u64 = 5360815; Min.u64 = 4; Max.u64 = 4; 
- node0.cpu0:decoder0.uop_delayed_rob_full.1 : Accumulator : Sum.u64 = 5227208; SumSQ.u64 = 5227208; Count.u64 = 5227208; Min.u64 = 1; Max.u64 = 1; 
- node0.cpu0:lsq.loads_issued.1 : Accumulator : Sum.u64 = 11409; SumSQ.u64 = 11409; Count.u64 = 11409; Min.u64 = 1; Max.u64 = 1; 
- node0.cpu0:lsq.stores_issued.1 : Accumulator : Sum.u64 = 7254; SumSQ.u64 = 7254; Count.u64 = 7254; Min.u64 = 1; Max.u64 = 1; 
- node0.cpu0:lsq.bytes_read.1 : Accumulator : Sum.u64 = 74746; SumSQ.u64 = 567212; Count.u64 = 11267; Min.u64 = 1; Max.u64 = 8; 
- node0.cpu0:lsq.bytes_stored.1 : Accumulator : Sum.u64 = 45378; SumSQ.u64 = 339102; Count.u64 = 6933; Min.u64 = 1; Max.u64 = 8; 
- node0.cpu0:lsq.fences_issued.1 : Accumulator : Sum.u64 = 1864; SumSQ.u64 = 1864; Count.u64 = 1864; Min.u64 = 1; Max.u64 = 1; 
- node0.cpu0:lsq.loads_executed.1 : Accumulator : Sum.u64 = 10283; SumSQ.u64 = 10283; Count.u64 = 10283; Min.u64 = 1; Max.u64 = 1; 
- node0.cpu0:lsq.stores_executed.1 : Accumulator : Sum.u64 = 6676; SumSQ.u64 = 6676; Count.u64 = 6676; Min.u64 = 1; Max.u64 = 1; 
- node0.cpu0:lsq.fences_executed.1 : Accumulator : Sum.u64 = 1296; SumSQ.u64 = 1296; Count.u64 = 1296; Min.u64 = 1; Max.u64 = 1; 
- node0.cpu0:lsq.store_buffer_entries.1 : Accumulator : Sum.u64 = 174594; SumSQ.u64 = 781180; Count.u64 = 2015219; Min.u64 = 0; Max.u64 = 8; 
- node0.cpu0:lsq.stores_in_flight.1 : Accumulator : Sum.u64 = 2542717; SumSQ.u64 = 99743029; Count.u64 = 2015219; Min.u64 = 0; Max.u64 = 131; 
- node0.cpu0:lsq.loads_in_flight.1 : Accumulator : Sum.u64 = 909879; SumSQ.u64 = 1599317; Count.u64 = 2015219; Min.u64 = 0; Max.u64 = 13; 
- node0.cpu0:lsq.operations_pending : Accumulator : Sum.u64 = 267755; SumSQ.u64 = 1199631; Count.u64 = 2015219; Min.u64 = 0; Max.u64 = 11; 
- node0.cpu0:rocc.roccs_issued.1 : Accumulator : Sum.u64 = 13; SumSQ.u64 = 13; Count.u64 = 13; Min.u64 = 1; Max.u64 = 1; 
- node0.cpu0.instructions_retired.1 : Accumulator : Sum.u64 = 50291; SumSQ.u64 = 102665; Count.u64 = 2015219; Min.u64 = 0; Max.u64 = 4; 
- node0.cpu0.instructions_decoded.1 : Accumulator : Sum.u64 = 162743; SumSQ.u64 = 1033409; Count.u64 = 2015219; Min.u64 = 0; Max.u64 = 12; 
- node0.cpu0.instructions_issued.1 : Accumulator : Sum.u64 = 62660; SumSQ.u64 = 135340; Count.u64 = 2015219; Min.u64 = 0; Max.u64 = 4; 
- node0.cpu0.loads_issued.1 : Accumulator : Sum.u64 = 11409; SumSQ.u64 = 11409; Count.u64 = 11409; Min.u64 = 1; Max.u64 = 1; 
- node0.cpu0.stores_issued.1 : Accumulator : Sum.u64 = 7254; SumSQ.u64 = 7254; Count.u64 = 7254; Min.u64 = 1; Max.u64 = 1; 
- node0.cpu0.branch_mispredicts.1 : Accumulator : Sum.u64 = 2340; SumSQ.u64 = 2340; Count.u64 = 2340; Min.u64 = 1; Max.u64 = 1; 
- node0.cpu0.branches.1 : Accumulator : Sum.u64 = 11798; SumSQ.u64 = 11798; Count.u64 = 11798; Min.u64 = 1; Max.u64 = 1; 
- node0.cpu0.cycles.1 : Accumulator : Sum.u64 = 2015219; SumSQ.u64 = 2015219; Count.u64 = 2015219; Min.u64 = 1; Max.u64 = 1; 
- node0.cpu0.rob_slots_in_use.1 : Accumulator : Sum.u64 = 47676864; SumSQ.u64 = 2866176488; Count.u64 = 2015219; Min.u64 = 0; Max.u64 = 63; 
- node0.cpu0.rob_cleared_entries.1 : Accumulator : Sum.u64 = 112427; SumSQ.u64 = 6284811; Count.u64 = 2340; Min.u64 = 0; Max.u64 = 62; 
- node0.cpu0.syscall-cycles.1 : Accumulator : Sum.u64 = 26872; SumSQ.u64 = 26872; Count.u64 = 26872; Min.u64 = 1; Max.u64 = 1; 
- node0.cpu0.phys_int_reg_in_use.1 : Accumulator : Sum.u64 = 74572603; SumSQ.u64 = 2775950279; Count.u64 = 2015219; Min.u64 = 35; Max.u64 = 51; 
- node0.cpu0.phys_fp_reg_in_use.1 : Accumulator : Sum.u64 = 66502227; SumSQ.u64 = 2194573491; Count.u64 = 2015219; Min.u64 = 33; Max.u64 = 33; 
-Simulation is complete, simulated time: 876.62 us
