--- src/sst/elements/vanadis/tests/small/basic-io/hello-world-cpp-rocc/riscv64/vanadis.stdout.gold
@@ -1,1 +1,0 @@
-Hello World from C++
