+++ src/sst/elements/vanadis/tests/small/basic-math/sqrt-float/riscv64/sst.stdout.gold
@@ -24,7 +24,7 @@ all process have exited
  node0.cpu0:lsq.operations_pending : Accumulator : Sum.u64 = 54146; SumSQ.u64 = 156298; Count.u64 = 24162323; Min.u64 = 0; Max.u64 = 9; 
  node0.cpu0.instructions_retired.1 : Accumulator : Sum.u64 = 3980010; SumSQ.u64 = 6014486; Count.u64 = 24162323; Min.u64 = 0; Max.u64 = 4; 
  node0.cpu0.instructions_decoded.1 : Accumulator : Sum.u64 = 4415947; SumSQ.u64 = 9396271; Count.u64 = 24162323; Min.u64 = 0; Max.u64 = 12; 
- node0.cpu0.instructions_issued.1 : Accumulator : Sum.u64 = 4053758; SumSQ.u64 = 4250988; Count.u64 = 24162323; Min.u64 = 0; Max.u64 = 4; 
+ node0.cpu0.instructions_issued.1 : Accumulator : Sum.u64 = 4053735; SumSQ.u64 = 4250965; Count.u64 = 24162323; Min.u64 = 0; Max.u64 = 4; 
  node0.cpu0.loads_issued.1 : Accumulator : Sum.u64 = 22647; SumSQ.u64 = 22647; Count.u64 = 22647; Min.u64 = 1; Max.u64 = 1; 
  node0.cpu0.stores_issued.1 : Accumulator : Sum.u64 = 11079; SumSQ.u64 = 11079; Count.u64 = 11079; Min.u64 = 1; Max.u64 = 1; 
  node0.cpu0.branch_mispredicts.1 : Accumulator : Sum.u64 = 10004; SumSQ.u64 = 10004; Count.u64 = 10004; Min.u64 = 1; Max.u64 = 1;
