+++ src/sst/elements/vanadis/tests/small/rocc/basic-rocc/Makefile
@@ -1,0 +1,17 @@
+
+CC=$(ARCH)-linux-musl-gcc
+CXX=$(ARCH)-linux-musl-g++
+
+CFLAGS=-O3
+CXXFLAGS=-O3
+LDFLAGS=-static
+
+PROG=basic-rocc
+
+$(PROG) : $(PROG).c
+	$(CC) $(CFLAGS) $(LDFLAGS) -o $(ARCH)/$@ $<
+
+clean:
+	rm -r $(ARCH)/$(PROG)
+
+
