+++ src/sst/elements/vanadis/tests/small/rocc/basic-rocc/riscv64/2rocc/vanadis.stdout.gold
@@ -1,0 +1,8 @@
+Result from RoCC 0: 2147482888
+Result from RoCC 0: 89047
+Loaded value 1234 in RoCC 0.
+Stored value 1234 from RoCC 0.
+Result from RoCC 1: 89104
+Result from RoCC 1: 6325892
+Loaded value 4321 in RoCC 1.
+Stored value 4321 from RoCC 1.
