+++ src/sst/elements/vanadis/tests/testsuite_default_vanadis.py
@@ -91,6 +91,13 @@ def build_vanadis_test_matrix():
             testlist.append(["basic_vanadis.py", location, test,arch, 1,32, "32thread", 300])
             testlist.append(["basic_vanadis.py", location, test,arch, 4,8, "4core-8thread", 300])
 
+    location="small/rocc"
+    tests = ['basic-rocc']
+    arch_list = ["riscv64"]
+    for test in tests:
+        for arch in arch_list:
+            testlist.append(["rocc_vanadis.py", location, test,arch, 1,1, "2rocc", 300])
+
     # Process each line and crack up into an index, hash, options and sdl file
     for testnum, test_info in enumerate(testlist):
         # Make testnum start at 1
