+++ src/sst/elements/vanadis/vanadis.h
@@ -180,10 +180,7 @@ public:
     // Optional since there is nothing to document
     SST_ELI_DOCUMENT_SUBCOMPONENT_SLOTS(
         { "lsq", "Load-Store Queue for Memory Access", "SST::Vanadis::VanadisLoadStoreQueue" },
-        { "rocc0", "RoCC accelerator interface", "SST::Vanadis::VanadisRoCCInterface" },
-        { "rocc1", "RoCC accelerator interface", "SST::Vanadis::VanadisRoCCInterface" },
-        { "rocc2", "RoCC accelerator interface", "SST::Vanadis::VanadisRoCCInterface" },
-        { "rocc3", "RoCC accelerator interface", "SST::Vanadis::VanadisRoCCInterface" },
+        { "rocc", "RoCC accelerator interface(s)", "SST::Vanadis::VanadisRoCCInterface" },
         { "mem_interface_inst", "Interface to memory system for instructions", "SST::Interfaces::StandardMem" },
     )
 
@@ -303,10 +300,6 @@ private:
     std::vector<VanadisDecoder*>                            thread_decoders;
     std::vector<const VanadisDecoderOptions*>               isa_options;
 
-    std::deque<VanadisInstruction*> rocc0_queue;
-    std::deque<VanadisInstruction*> rocc1_queue;
-    std::deque<VanadisInstruction*> rocc2_queue;
-    std::deque<VanadisInstruction*> rocc3_queue;
 
     std::vector<VanadisFunctionalUnit*> fu_int_arith;
     std::vector<VanadisFunctionalUnit*> fu_int_div;
@@ -331,14 +324,8 @@ private:
     VanadisLoadStoreQueue* lsq;
     StandardMem*           memInstInterface;
 
-    VanadisRoCCInterface* rocc0;
-    bool has_rocc0;
-    VanadisRoCCInterface* rocc1;
-    bool has_rocc1;
-    VanadisRoCCInterface* rocc2;
-    bool has_rocc2;
-    VanadisRoCCInterface* rocc3;
-    bool has_rocc3;
+    std::vector<VanadisRoCCInterface*> roccs_;
+    std::vector<std::deque<VanadisInstruction*>> rocc_queues_;
 
     bool* halted_masks;
     bool  print_int_reg;
